Move live density report stack under app.core.reports.density.
Issue #798 Phase 6 renames misleading new_* modules while keeping temporary
DeprecationWarning shims and a façade alias for callers.
