+++ app/core/reports/__init__.py
@@ -1,0 +1,1 @@
+"""Report generation packages (Markdown / operational reports)."""
